Small simulated item pool example.

A data.frame containing 100 not yet calibrated items with different categorical and metric properties.

Usage

items_pilot

Format

A data.frame .

item

Item identifier.

diffCategory

Item difficulty (five categories).

format

Item format (multiple choice, constructed multiple choice, or open answer).

domain

Item domain (listening, reading, or writing).

time

Average response times in seconds.

exclusions

Items which can not be in the same test form.